+++ src/Tabby/Asset/AssetBase.cpp
@@ -1,6 +1,0 @@
-#include <Tabby/Asset/AssetBase.h>
-#include <Tabby/Asset/AssetManager.h>
-
-namespace Tabby {
-
-}