+++ src/Tabby/Asset/AssetManager.cpp
@@ -3,6 +3,7 @@
 #include <Tabby/Asset/AssetManager.h>
 #include <Tabby/Renderer/Font.h>
 #include <Tabby/Renderer/Texture.h>
+#include <Tabby/Audio/Audio.h>
 
 #include <Tabby/Utils/Utils.h>
 #include <stb_image.h>
@@ -11,25 +12,11 @@ namespace Tabby {
 
 AssetManager::AssetManager()
 {
-    TB_CORE_ASSERT_TAGGED(!s_Instance, "AssetManager already Initialized!");
-    s_Instance = this;
-
-    TB_CORE_INFO("Initialized asset manager");
 }
 
 AssetManager::~AssetManager()
 {
     FullUnload();
-}
-
-void AssetManager::Init()
-{
-    s_Instance = new AssetManager;
-}
-
-void AssetManager::Shutdown()
-{
-    delete s_Instance;
 }
 
 void AssetManager::FullUnload()
@@ -50,7 +37,7 @@ AssetHandle AssetManager::LoadAssetSource(std::filesystem::path path, const Asse
     case AssetType::MATERIAL:
         break;
     case AssetType::AUDIO_SRC:
-        break;
+        return ImportAudioSource(path, id);
     case AssetType::IMAGE_SRC:
         return ImportImageSource(path, id);
     case AssetType::FONT_SRC:
@@ -92,8 +79,7 @@ AssetHandle AssetManager::ImportImageSource(std::filesystem::path path, AssetHan
     {
 
         stbi_set_flip_vertically_on_load(true);
-        TB_PROFILE_SCOPE_NAME("stbi_load - AssetManager::ImportTexture");
-#if defined(TB_PLATFORM_ANDROID)
+        TB_PROFILE_SCOPE_NAME("stbi_load - AssetManager::ImportTextureSource");
         std::vector<unsigned char> imageData;
 
         SDL_RWops* rw = SDL_RWFromFile(path.c_str(), "rb");
@@ -114,19 +100,13 @@ AssetHandle AssetManager::ImportImageSource(std::filesystem::path path, AssetHan
         } else {
             // Handle file open error
             TB_CORE_ERROR("Could not open file {0}", path);
-            TB_CORE_INFO("Current working dir: {0}", std::filesystem::current_path());
         }
 
-        data.Data = stbi_load_from_memory(imageData.data(), static_cast<int>(imageData.size()), &width, &height, &channels, 4);
-        channels = 4;
-#else
-        data.Data = stbi_load(path.string().c_str(), &image_width, &image_height, &channels, 4);
-#endif
-
+        data.Data = stbi_load_from_memory(imageData.data(), static_cast<int>(imageData.size()), &image_width, &image_height, &channels, 4);
         channels = 4;
 
         if (data.Data == nullptr) {
-            TB_CORE_ERROR("TextureImporter::ImportTexture - Could not load texture from filepath: {}", path.string());
+            TB_CORE_ERROR("TextureImporter::ImportTextureSource - Could not load texture from filepath: {}", path.string());
         }
     }
 
@@ -143,14 +123,7 @@ AssetHandle AssetManager::ImportImageSource(std::filesystem::path path, AssetHan
     std::array<AssetFileSubresourceMetadata, 16> subresources_metadata = {};
 
     TextureSpecification texture_spec = {};
-    // switch (channels) {
-    // case 3:
-    //     texture_spec.Format = ImageFormat::RGB8;
-    //     break;
-    // case 4:
     texture_spec.Format = ImageFormat::RGBA8;
-    //     break;
-    // }
     texture_spec.Width = image_width;
     texture_spec.Height = image_height;
     texture_spec.type = ImageType::TYPE_2D;
@@ -171,57 +144,93 @@ AssetHandle AssetManager::ImportImageSource(std::filesystem::path path, AssetHan
 AssetHandle AssetManager::ImportFontSource(std::filesystem::path path, AssetHandle handle)
 {
 
-    //     if (m_UUIDs.find(path.string()) != m_UUIDs.end())
-    //         return m_UUIDs.at(path.string());
-    //
-    //     Buffer data;
-    //
-    //     {
-    //
-    //         TB_PROFILE_SCOPE_NAME("FT_New_Face - AssetManager::ImportTexture");
-    // #if defined(TB_PLATFORM_ANDROID)
-    //         std::vector<unsigned char> fontData;
-    //
-    //         SDL_RWops* rw = SDL_RWFromFile(path.c_str(), "rb");
-    //         if (rw != nullptr) {
-    //             Sint64 size = SDL_RWsize(rw);
-    //
-    //             if (size > 0) {
-    //                 fontData.resize(size);
-    //                 Sint64 bytesRead = SDL_RWread(rw, fontData.data(), 1, size);
-    //                 if (bytesRead != size) {
-    //                     // Handle read error
-    //                     TB_CORE_ERROR("Error reading file {0}", path);
-    //                     fontData.clear(); // Clear the imageData to indicate an error
-    //                 }
-    //             }
-    //
-    //             SDL_RWclose(rw);
-    //         } else {
-    //             // Handle file open error
-    //             TB_CORE_ERROR("Could not open file {0}", path);
-    //             TB_CORE_INFO("Current working dir: {0}", std::filesystem::current_path());
-    //         }
-    //
-    //         FT_New_Memory_Face(m_FTLibrary, fontData.data(), static_cast<int>(fontData.size()), 0, data.As<FT_Face>());
-    // #else
-    //         FT_New_Face(m_FTLibrary, path.string().c_str(), 0, data.As<FT_Face>());
-    // #endif
-    //
-    //         if (data.Data == nullptr) {
-    //             TB_CORE_ERROR("TextureImporter::ImportTexture - Could not load texture from filepath: {}", path.string());
-    //         }
-    //     }
-    //
-    //     FontSpecification fontSpecification = { path };
-    //
-    //     Shared<Font> font = CreateShared<Font>(fontSpecification, handle, data);
-    //     data.Release();
-    //
-    //     m_AssetRegistry.emplace(font->Handle, font);
-    //     m_UUIDs.emplace(path.string(), font->Handle);
-    //
-    //     return font->Handle;
+    if (m_UUIDs.find(path.string()) != m_UUIDs.end())
+        return m_UUIDs.at(path.string());
+
+    Buffer data;
+
+    {
+        TB_PROFILE_SCOPE_NAME("AssetManager::ImportFontSource");
+
+        SDL_RWops* rw = SDL_RWFromFile(path.c_str(), "rb");
+        if (rw != nullptr) {
+            Sint64 size = SDL_RWsize(rw);
+
+            if (size > 0) {
+                data.Data = new uint8_t[size];
+                data.Size = size;
+
+                Sint64 bytesRead = SDL_RWread(rw, data.Data, 1, size);
+                if (bytesRead != size) {
+                    TB_CORE_ERROR("Error reading file {0}", path);
+                    data.Release();
+                }
+            }
+
+            SDL_RWclose(rw);
+        } else {
+            TB_CORE_ERROR("Could not open file {0}", path);
+        }
+
+        if (data.Data == nullptr) {
+            TB_CORE_ERROR("AssetManager::ImportFontSource - Could not load font from filepath: {}", path.string());
+        }
+    }
+
+    Shared<Font> font = CreateShared<Font>(path, handle, data);
+
+    data.Release();
+
+    m_AssetRegistry.emplace(font->Handle, font);
+    m_UUIDs.emplace(path.string(), font->Handle);
+
+    return font->Handle;
+}
+
+AssetHandle AssetManager::ImportAudioSource(std::filesystem::path path, AssetHandle handle)
+{
+
+    if (m_UUIDs.find(path.string()) != m_UUIDs.end())
+        return m_UUIDs.at(path.string());
+
+    Buffer data;
+
+    {
+        TB_PROFILE_SCOPE_NAME("AssetManager::ImportAudioSource");
+
+        SDL_RWops* rw = SDL_RWFromFile(path.c_str(), "rb");
+        if (rw != nullptr) {
+            Sint64 size = SDL_RWsize(rw);
+
+            if (size > 0) {
+                data.Data = new uint8_t[size];
+                data.Size = size;
+
+                Sint64 bytesRead = SDL_RWread(rw, data.Data, 1, size);
+                if (bytesRead != size) {
+                    TB_CORE_ERROR("Error reading file {0}", path);
+                    data.Release();
+                }
+            }
+
+            SDL_RWclose(rw);
+        } else {
+            TB_CORE_ERROR("Could not open file {0}", path);
+        }
+
+        if (data.Data == nullptr) {
+            TB_CORE_ERROR("AssetManager::ImportAudioSource - Could not load audio from filepath: {}", path.string());
+        }
+    }
+
+    Shared<Audio> audio = CreateShared<Audio>(path, handle, data);
+
+    // data.Release();
+
+    m_AssetRegistry.emplace(audio->Handle, audio);
+    m_UUIDs.emplace(path.string(), audio->Handle);
+
+    return audio->Handle;
 }
 
 }