+++ src/Tabby/Foundation/Types.h
@@ -221,7 +221,8 @@ struct Buffer {
 
     void Release()
     {
-        free(Data);
+        if (Data)
+            free(Data);
         Data = nullptr;
         Size = 0;
     }