+++ src/Tabby/Renderer/Texture.cpp
@@ -26,21 +26,4 @@ Shared<Texture> Texture::Create(const TextureSpecification& specification, Asset
     TB_CORE_ASSERT_TAGGED(false, "Unknown RendererAPI!");
     return nullptr;
 }
-
-// Shared<Texture> Texture::Create(const std::string& path)
-// {
-//     switch (Renderer::GetAPI()) {
-//     case RendererAPI::API::None:
-//         TB_CORE_ASSERT_TAGGED(false, "RendererAPI::None is currently not supported!");
-//         return nullptr;
-//     case RendererAPI::API::OpenGL33:
-//         return CreateShared<OpenGL33Texture>(path);
-//     case RendererAPI::API::OpenGLES3:
-//         return CreateShared<OpenGLES3Texture>(path);
-//     }
-//
-//     TB_CORE_ASSERT_TAGGED(false, "Unknown RendererAPI!");
-//     return nullptr;
-// }
-
 }